+++ packages/lib/src/typeChecking/arraySet.ts
@@ -1,3 +1,4 @@
+import { getGlobalConfig } from "../globalConfig/globalConfig"
 import { modelTypeKey } from "../model/metadata"
 import { modelInfoByClass } from "../modelShared/modelInfo"
 import { isObject } from "../utils"
@@ -80,6 +81,19 @@ export function typesArraySet<T extends AnyType>(
           [modelTypeKey]: modelInfo.name,
           items: sn.items.map((v) => valueChecker.fromSnapshotProcessor(v)),
         }
+      },
+
+      (sn: { items: unknown[]; [modelTypeKey]?: string }) => {
+        const snCopy = {
+          ...sn,
+          items: sn.items.map((v) => valueChecker.toSnapshotProcessor(v)),
+        }
+
+        if (getGlobalConfig().avoidModelTypeInTypedModelSnapshotsIfPossible) {
+          delete snCopy[modelTypeKey]
+        }
+
+        return snCopy
       }
     )